requirements.txt
streamlit==1.41.1
groq==0.15.0
google-genai==1.5.0
reportlab==4.2.5
google-api-python-client==2.159.0
google-auth==2.37.0
